+++ changes.md
@@ -1,3 +1,7 @@
+#### 1.2.1
+* Fixed Content-Type response headers.
+* Fixed dependency saving/updating, which did not work when forms were saved via AJAX.
+
 #### 1.2.0
 * Added bundled vendor dependencies.
 * Added changes.md file.